fix: npm dependencies and working dev environment
