+++ pages/_app.js
@@ -1,3 +1,4 @@
+// pages/_app.js
 import "../styles/globals.css";
 
 export default function App({ Component, pageProps }) {
